refactor: consolidate storage and remove redundant scaffolding
