+++ examples/providers/test/live-smoke.test.ts
@@ -6,9 +6,7 @@ import {
   PHASE7_LIVE_GATE_ENV,
   phase1Trip,
   phase7LiveProfileEnabled,
-  phase7TravelPlannerProfile,
   TravelPlan,
-  TravelPlannerPhase7Profile,
   TravelPlannerRuntimeLayer,
 } from "@effect-agent/testing/TravelPlanner";
 import { OpenAiClient } from "@effect/ai-openai";
@@ -18,38 +16,9 @@ import { describe, expect, it } from "vite-plus/test";
 
 import { OpenAiTravelPlanner } from "../src/index.ts";
 
-// ---------------------------------------------------------------------------
-// The P7 test-side live-gating pattern (plan §6): ONE environment predicate,
-// exported by the fixture, applied with `describe.skipIf`. The ungated
-// describe below runs on every ordinary `bun run test`; the live describe is
-// skipped unless BOTH the explicit opt-in flag and the credential are present,
-// so no ordinary gate ever makes a network request.
-// ---------------------------------------------------------------------------
-
 const liveEnabled = phase7LiveProfileEnabled(process.env);
 
-describe("TEST-014 P7 Travel Planner dual-profile pin", () => {
-  it("pins the deterministic offline conformance profile alongside the opt-in live-model profile", () => {
-    const decoded = Schema.decodeUnknownSync(TravelPlannerPhase7Profile)(
-      Schema.encodeSync(TravelPlannerPhase7Profile)(phase7TravelPlannerProfile),
-    );
-
-    expect(decoded).toEqual(phase7TravelPlannerProfile);
-    expect(phase7TravelPlannerProfile).toEqual({
-      phase: "P7",
-      offlineConformanceDeterministic: true,
-      offlineRequiresCredentials: false,
-      liveProfileOptIn: true,
-      // Live profiles bind live MODEL Layers over the SAME deterministic
-      // suppliers; "selected supplier Layers" is honestly scoped to
-      // live-model-only (P7 plan decision 9).
-      liveModelLayers: true,
-      liveSupplierLayers: false,
-      structurallyRedactedTranscripts: true,
-      exactlyOnceExternalEffects: false,
-    });
-  });
-
+describe("Travel Planner live-model opt-in", () => {
   it("enables the live profile only for EFFECT_AGENT_LIVE=1 plus a present credential", () => {
     expect(PHASE7_LIVE_GATE_ENV).toBe("EFFECT_AGENT_LIVE");
     expect(PHASE7_LIVE_CREDENTIAL_ENV).toBe("OPENAI_API_KEY");
@@ -64,14 +33,7 @@ describe("TEST-014 P7 Travel Planner dual-profile pin", () => {
   });
 });
 
-// ---------------------------------------------------------------------------
-// Opt-in live smoke (class E, one Run, rate-limited by construction): the
-// existing `OpenAiTravelPlanner` binding over the application-supplied OpenAI
-// client Layer and the SAME deterministic supplier/travel-service Layers the
-// offline suites use. Transcript evidence never leaves this test in raw form:
-// every RunEvent is Schema-encoded and passed through the structural Redactor
-// before it can appear in an assertion message (SEC-008, testing.md §12).
-// ---------------------------------------------------------------------------
+// Live model over deterministic suppliers; transcript assertions use structural redaction.
 
 const OpenAiClientLayer = OpenAiClient.layerConfig({
   apiKey: Config.redacted(PHASE7_LIVE_CREDENTIAL_ENV),
@@ -85,7 +47,7 @@ const liveSmokeLayer = Layer.mergeAll(
 
 const encodeRunEvent = Schema.encodeEffect(RunEvent);
 
-describe.skipIf(!liveEnabled)("TEST-014 P7 Travel Planner live-model smoke (opt-in)", () => {
+describe.skipIf(!liveEnabled)("Travel Planner live-model smoke (opt-in)", () => {
   it(
     "completes one live planning Run over deterministic suppliers and emits only structurally redacted transcript evidence",
     { timeout: 120_000 },
@@ -129,8 +91,7 @@ describe.skipIf(!liveEnabled)("TEST-014 P7 Travel Planner live-model smoke (opt-
             }
           }
 
-          // The live Run exercised the deterministic supplier Layers (the
-          // supplier side stays offline by design — decision 9).
+          // The live Run exercised the deterministic supplier Layers.
           const toolResults = events.filter((event) => event._tag === "ToolCallSucceeded");
 
           expect(toolResults.length).toBeGreaterThan(0);
