+++ examples/repo-ops/package.json
@@ -14,9 +14,7 @@
     "@effect-agent/platform-node": "workspace:*",
     "@effect-agent/sandbox": "workspace:*",
     "@effect-agent/sandbox-local": "workspace:*",
-    "@effect-agent/testing": "workspace:*",
     "@effect-agent/thread": "workspace:*",
-    "@effect/ai-openai": "catalog:",
     "effect": "catalog:"
   },
   "devDependencies": {
